+++ crates/cli/Cargo.toml
@@ -10,6 +10,7 @@ tokio = { version = "1.17.0", features = ["macros"] }
 clap = { version = "4.0.32", features = ["derive"] }
 serde = { version = "1.0.152", features =["derive"] }
 clairvoyance = { path = "../clairvoyance" }
+utils = { path = "../utils" }
 thiserror = "1.0.38"
 ethers = "1.0.0"
 eyre = "0.6.6"
