+++ crates/utils/src/lib.rs
@@ -1,2 +1,2 @@
 pub mod tokens;
-pub mod utils;
+pub mod chain_tools;
